# api/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  model.PostInput:
    properties:
      categoryId:
        type: integer
      contentMd:
        type: string
      coverUrl:
        type: string
      excerpt:
        type: string
      kind:
        type: string
      pinned:
        type: boolean
      slug:
        type: string
      status:
        type: string
      tagIds:
        items:
          type: integer
        type: array
      title:
        type: string
      version:
        type: integer
    type: object
  model.Revision:
    properties:
      createdAt:
        type: string
      id:
        type: integer
      snapshot:
        $ref: '#/definitions/model.PostInput'
    type: object
...
  model.WritingDraft:
    properties:
      key:
        type: string
      snapshot:
        $ref: '#/definitions/model.PostInput'
      updatedAt:
        type: string
      version:
        type: integer
    type: object
info:
  contact: {}
  description: 博客 API。登录使用同源 HTTPS HttpOnly Cookie，前台和后台会话独立。先调用对应登录接口，再使用 Try it
    out；不使用 Bearer Token。当前文档覆盖健康检查、登录会话、文章及写作工作流。
  title: Kakozane Blog API
  version: "1.0"
paths:
...
  /admin/drafts/{key}:
    delete:
      parameters:
      - description: post-ID、note-ID、post-new 或 note-new
        in: path
        name: key
        required: true
        type: string
      - description: 当前草稿版本
        in: query
        name: version
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: 操作成功
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
        "503":
          description: Service Unavailable
          schema:
            additionalProperties:
              type: string
            type: object
      summary: DeleteDraft（需后台登录）
      tags:
      - 草稿与版本
    get:
      parameters:
      - description: post-ID、note-ID、post-new 或 note-new
        in: path
        name: key
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.WritingDraft'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
        "503":
          description: Service Unavailable
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Draft（需后台登录）
      tags:
      - 草稿与版本
    put:
      consumes:
      - application/json
      parameters:
      - description: post-ID、note-ID、post-new 或 note-new
        in: path
        name: key
        required: true
        type: string
      - description: 携带当前草稿 version，首次为 0
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.WritingDraft'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.WritingDraft'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
        "503":
          description: Service Unavailable
          schema:
            additionalProperties:
              type: string
            type: object
      summary: SaveDraft（需后台登录）
      tags:
      - 草稿与版本
...
  /admin/posts/{id}/purge:
    delete:
      parameters:
      - description: 文章 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: 操作成功
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
        "503":
          description: Service Unavailable
          schema:
            additionalProperties:
              type: string
            type: object
      summary: PurgePost（需后台登录）
      tags:
      - 草稿与版本
  /admin/posts/{id}/restore:
    post:
      parameters:
      - description: 文章 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: 操作成功
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
        "503":
          description: Service Unavailable
          schema:
            additionalProperties:
              type: string
            type: object
      summary: RestorePost（需后台登录）
      tags:
      - 草稿与版本
  /admin/posts/{id}/revisions:
    get:
      parameters:
      - description: 文章 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              items:
                items:
                  $ref: '#/definitions/model.Revision'
                type: array
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
        "503":
          description: Service Unavailable
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Revisions（需后台登录）
      tags:
      - 草稿与版本
...
schemes:
- https
swagger: "2.0"